requirements.txt
aiosqlite==0.22.1
anthropic==0.86.0
cryptography==46.0.5
fastapi==0.135.2
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-timeout==2.4.0
SQLAlchemy==2.0.48
starlette==1.0.0
websockets==16.0
